fix(platform): resolve p1 review issues for unify-chat-threads
- add integration test for scoped (?agentId=) path returning agent field
- reference tracking issue #10284 in agentId deprecation JSDoc comments
- extract useChatThreadsTitleLabels hook to zero-sidebar-shared to eliminate duplication between sidebar-threads and zero-chat-list-page
