+++ turbo/packages/core/src/contracts/chat-threads.ts
@@ -34,15 +34,15 @@ const chatThreadListItemSchema = z.object({
   id: z.string(),
   title: z.string().nullable(),
   /**
-   * Deprecated: will be removed once every consumer reads `agent.id`.
-   * Kept temporarily so existing fixtures still parse while the unified-list
-   * feature switch rolls out.
+   * @deprecated Use `agent.id` instead. Will be removed in #10284 once every
+   * consumer reads `agent.id` and the UnifyChatThreads flag has fully rolled out.
+   * Kept temporarily so existing fixtures still parse during the rollout window.
    */
   agentId: z.string(),
   /**
    * Owning agent snapshot. Always emitted by the server; kept optional on the
    * schema so older fixtures that predate the unified-list rollout still
-   * validate until they are migrated.
+   * validate until they are migrated (tracked in #10284).
    */
   agent: z
     .object({
